Improved persisted draft cleanup
