+++ apps/lite/ui/src/commit.ts
@@ -1,4 +1,27 @@
 import type { Commit, ForgeInfo } from "@gitbutler/but-sdk";
+import { queryOptions, useMutation } from "@tanstack/react-query";
+import * as idb from "idb-keyval";
+
+const draftCommitMessageKey = (projectId: string): string => `commit_message_draft:v1:${projectId}`;
+
+export const draftCommitMessageQueryOptions = (projectId: string) =>
+	queryOptions({
+		queryKey: [projectId, "commitMessageDraft"],
+		queryFn: async () => (await idb.get<string>(draftCommitMessageKey(projectId))) ?? "",
+	});
+
+export const usePersistDraftCommitMessage = () =>
+	useMutation({
+		mutationFn: ({ projectId, message }: { projectId: string; message: string }) =>
+			message === ""
+				? idb.del(draftCommitMessageKey(projectId))
+				: idb.set(draftCommitMessageKey(projectId), message),
+		onSuccess: (_data, input, _res, ctx) =>
+			ctx.client.setQueryData(
+				draftCommitMessageQueryOptions(input.projectId).queryKey,
+				input.message,
+			),
+	});
 
 export const shortCommitId = (commitId: string): string => commitId.slice(0, 7);
